+++ docs/CHANGE_LOG.md
@@ -2,6 +2,9 @@
 
 ## Version 0.2.0:
 
+#### **Public Functions:**
+- Added public function "get()" now has an extra parameter called "includeChildren" (defaults to true), which states if the children for an element should be included.
+
 #### **Configuration Options:**
 - Added new configuration option "formattingNodeTypes", which states the text formatting nodes to check for (defaults to "b strong i em mark small del ins sub sup").